+++ src/js/download_adapter.jquery.js
@@ -22,12 +22,12 @@
  *
  * @author Artur Barseghyan (https://github.com/barseghyanartur)
  * @url https://github.com/barseghyanartur/prismjs-sphinx
- * @version 1.1.2
+ * @version 1.2.0
  */
 
 $(document).ready(function() {
     // Find all download links by their class
-    $(".prismjs-sphinx a.reference.download.internal").each(function(index) {
+    $(".prismjs-sphinx-download a.reference.download.internal").each(function(index) {
         // Create a unique id for the additional content div
         let contentID = 'additional-content-' + index;